Update Multitrait.R
Changing initalization of factors in setCov.FA  This new approach allows for the user to provide inital values for the factor loadings ($W). This serves to purposes: (i) it can be used to avoid numerical problems when factanal() does not converge, and (ii) it could be used to set factor scales. For instance if $W[i,j]=1 and $M[i,j]=FALSE, the loading of trait 1 on factor 1 is 1 and this won't be updated.